✨ feat(editor): '+' new-file action in the tab bar (#612)
Problem: creating a file required reaching for the left-panel "New File"
button; there was no affordance in the editor tab bar itself.

Fix: add an optional `onNewFile` prop to WorkspaceShell, surfaced as a "+"
button in each group's pinned tab-bar action cluster (rendered only when the
host wires the callback). The app passes `onNewFile` from StrudelEditorClient.

To keep the "+" byte-for-byte identical to the left-panel button, the new-file
flow (prompt copy + supported-extension validation + unknown-extension toast)
is extracted into a shared `lib/newFile.ts` (`promptAndCreateFile` +
`extensionToLanguage`). FileTree now delegates `handleNewFile` to it and imports
`extensionToLanguage` (its drag-drop import path) from the same single source —
removing the duplicated extension table.
